Spreadsheet Verification
Various changes made to main code to allow for specific testing procedures and printouts for spreadsheet verification.

+++ Capstone_Spreadsheet_Verification/Pot.h
@@ -2,7 +2,7 @@ class Pot {
   private:
 
   //Class Variables
-  int _Product;
+  double _Product;
   int _scale = 10; //Scales the potentiometer reading by this factor ie. 1024 divide by ____ 
 
   public:
@@ -21,7 +21,7 @@ class Pot {
    * low = the potentiometer value at min stroke
    * distance = the overall physical distance between the max and min stroke (input decides units)
    */
-  int Pot_Distance(char trigPin, int high, int low, int distance) {
+  double Pot_Distance(char trigPin, int high, int low, double distance) {
     _Product = analogRead(trigPin);
     _Product = _Product / _scale; //SCALING VALUE reducing the number of measurments from 1024 to ____
     _Product = (_Product-low)*distance/(high - low); //Converts the potentiometer reading into a distance
